pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vertical-fl"
version = "1.0.0"
description = "PyTorch Vertical FL with Flower"
license = "Apache-2.0"
dependencies = [
    "flwr[simulation]>=1.13.1",
    "flwr-datasets>=0.3.0",
    "numpy==1.24.4",
    "pandas==2.0.3",
    "scikit-learn==1.3.2",
    "torch==2.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.flwr.app]
publisher = "flwrlabs"

[tool.flwr.app.components]
serverapp = "vertical_fl.server_app:app"
clientapp = "vertical_fl.client_app:app"

[tool.flwr.app.config]
num-server-rounds = 3
learning-rate = 0.1

[tool.flwr.federations]
default = "local-simulation"

[tool.flwr.federations.local-simulation]
options.num-supernodes = 2 # Note that this example will require changes to how VFL is implemented
